fixed widget render error on signin page by moving localStorage.getItem to store.js
