+++ src/firebase/useAuth.jsx
@@ -3,30 +3,21 @@ import * as firebase from 'firebase/app';
 import { useDispatch } from 'react-redux';
 import { setUser } from '../redux/actions';
 
-function getUserFromLocalStorage() {
-  const userString = localStorage.getItem('user');
-  if (userString) {
-    const user = JSON.parse(userString);
-    return user;
-  }
-  return null;
-}
+export const USER_LOCAL_STORAGE_KEY = 'user';
 
 function useAuth() {
   const dispatch = useDispatch();
 
   React.useEffect(() => {
-    dispatch(setUser(getUserFromLocalStorage()));
-
     const unsubscribe = firebase
       .auth()
       .onAuthStateChanged((user) => {
         if (user) {
           dispatch(setUser(user));
-          localStorage.setItem('user', JSON.stringify(user));
+          localStorage.setItem(USER_LOCAL_STORAGE_KEY, JSON.stringify(user));
         } else {
           dispatch(setUser(null));
-          localStorage.removeItem('user');
+          localStorage.removeItem(USER_LOCAL_STORAGE_KEY);
         }
       });
